persistencia de usuario

+++ foi-top/domain/Item-coast-event-dto.go
@@ -6,7 +6,7 @@ type ItemCoastEventDTO struct {
 	Descricao string
 	Valor float32
 	Unidade int
-	EventoId  int
+	EventoId  int64
 
 }
