+++ foi-top/persistence/entitys/baseModel.go
@@ -4,7 +4,7 @@ import "time"
 
 // gorm.Model definition
 type BaseModel struct {
-	ID        int `gorm:"primary_key;AUTO_INCREMENT"`
+	ID        int64 `gorm:"primary_key;AUTO_INCREMENT"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	DeletedAt *time.Time
